ENH: Store progress in separate file
Upon unexpected termination, the config file may become corrupted as it is trying to store the last change. Therefore do not re-save the config file, but load it only upon startup.

To keep track of the last change, store it in a separate file.
If that becomes corrupted, OrthancForward can still run, and will only reprocess changes in the worst case.